pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-quest"
version = "0.1.0"
description = "Drop into any fiction. Solve code-backed puzzles. Command AI companions."
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "anthropic>=0.40",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-tmp-files>=0.0.2",
]

[project.scripts]
aq = "agentic_quest.cli:main"
engine = "agentic_quest.cli:engine"

[tool.setuptools.packages.find]
where = ["src"]
